Changing name on Controllers to singular. Adding Post and Put methods to PaymentTypeController
