+++ BangazonAPI/Controllers/PaymentTypeController.cs
@@ -12,11 +12,11 @@ namespace BangazonAPI.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class PaymentTypesControler : ControllerBase
+    public class PaymentTypeController : ControllerBase
     {
         private readonly IConfiguration _config;
 
-        public PaymentTypesControler(IConfiguration config)
+        public PaymentTypeController(IConfiguration config)
         {
             _config = config;
         }
@@ -40,7 +40,7 @@ namespace BangazonAPI.Controllers
                     cmd.CommandText = @"SELECT p.Id AS PayId,
                                         p.AcctNumber,
                                         p.Name,
-                                        p.CustomerId AS CusId
+                                        p.CustomerId AS CusId,
                                         c.FirstName,
                                         c.LastName 
                                         FROM PaymentType p
@@ -85,7 +85,7 @@ namespace BangazonAPI.Controllers
                     cmd.CommandText = @"SELECT p.Id AS PayId,
                                         p.AcctNumber,
                                         p.Name,
-                                        p.CustomerId AS CusId
+                                        p.CustomerId AS CusId,
                                         c.FirstName,
                                         c.LastName 
                                         FROM PaymentType p
@@ -120,20 +120,97 @@ namespace BangazonAPI.Controllers
 
         // POST: api/PaymentType
         [HttpPost]
-        public void Post([FromBody] string value)
+        public async Task<IActionResult> Post([FromBody] PaymentType paymentType)
         {
+            using (SqlConnection conn = Connection)
+            {
+                conn.Open();
+                using (SqlCommand cmd = conn.CreateCommand())
+                {
+                    cmd.CommandText = @"INSERT INTO PaymentType (AcctNumber, Name, CustomerId) 
+                                        OUTPUT INSERTED.Id
+                                        VALUES (@acctNumber, @name, @customerId)";
+                    cmd.Parameters.Add(new SqlParameter("@acctNumber", paymentType.AcctNumber));
+                    cmd.Parameters.Add(new SqlParameter("@name", paymentType.Name));
+                    cmd.Parameters.Add(new SqlParameter("@customerId", paymentType.CustomerId));
+
+                    paymentType.Id = (int)await cmd.ExecuteScalarAsync();
+
+                    return CreatedAtRoute("GetPaymentType", new { id = paymentType.Id }, paymentType);
+                }
+            }
         }
 
         // PUT: api/PaymentType/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] string value)
+        public async Task<IActionResult> Put([FromRoute] int id, [FromBody] PaymentType paymentType)
         {
+            try
+            {
+                using (SqlConnection conn = Connection)
+                {
+                    conn.Open();
+                    using (SqlCommand cmd = conn.CreateCommand())
+                    {
+                        cmd.CommandText = @"UPDATE PaymentType
+                                            SET AcctNumber = @acctNumber,
+                                                Name = @name,
+                                                CustomerId = @customerId,
+                                            WHERE Id = @id";
+                        cmd.Parameters.Add(new SqlParameter("@acctNumber", paymentType.AcctNumber));
+                        cmd.Parameters.Add(new SqlParameter("@name", paymentType.Name));
+                        cmd.Parameters.Add(new SqlParameter("@customerId", paymentType.CustomerId));
+                        cmd.Parameters.Add(new SqlParameter("@id", id));
+
+                        int rowsAffected = await cmd.ExecuteNonQueryAsync();
+                        if (rowsAffected > 0)
+                        {
+                            return new StatusCodeResult(StatusCodes.Status204NoContent);
+                        }
+                        throw new Exception("No rows affected");
+                    }
+                }
+            }
+            catch (Exception)
+            {
+                if (!PaymentTypeExists(id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
         }
 
         // DELETE: api/ApiWithActions/5
         [HttpDelete("{id}")]
         public void Delete(int id)
         {
+        }
+        private bool PaymentTypeExists(int id)
+        {
+            using (SqlConnection conn = Connection)
+            {
+                conn.Open();
+                using (SqlCommand cmd = conn.CreateCommand())
+                {
+                    cmd.CommandText = @"SELECT p.Id AS PayId,
+                                        p.AcctNumber,
+                                        p.Name,
+                                        p.CustomerId AS CusId,
+                                        c.FirstName,
+                                        c.LastName 
+                                        FROM PaymentType p
+                                        LEFT JOIN Customer c ON c.Id = p.CustomerId
+                                        WHERE p.Id = @id";
+                    cmd.Parameters.Add(new SqlParameter("@id", id));
+
+                    SqlDataReader reader = cmd.ExecuteReader();
+                    return reader.Read();
+                }
+            }
         }
     }
 }
